Ensure quaternion components are not 1.
Previously, we would ensure that both quanernions are normalized, and
that their components are NOT close to 1.
The code contained bugs however, and there were cases where quaternions
not meeting either of these conditions could be returned.
Making it so both of these are true is not trivial, and after doing some
research and testing, it seems quaternions being normalized is not that
important for our IK solution *. Quaternions not having components being
one is essential though, so norm_to_quat and tests have been updated to make
sure it's never the case.

Also, improve the way we check for validity of the quaternion by
checking the angle between the two vectors, rather than checking the
value of their components.

* not saying it has 0 impact though